# PERF: MultiIndex creation in explode