--- a/geopandas/geoseries.py
+++ b/geopandas/geoseries.py
@@ -812,13 +812,15 @@ class GeoSeries(GeoPandasBase, Series):
             # extract original index values based on integer index
             outer_index = self.index.take(outer_idx)
 
-            index = zip(outer_index, inner_index)
+            index_arrays = []
+            for lvl in range(outer_index.nlevels):
+                index_arrays.append(outer_index.get_level_values(lvl).values)
 
-            # if self.index is a MultiIndex then index is a list of nested tuples
-            if isinstance(self.index, MultiIndex):
-                index = [tuple(outer) + (inner,) for outer, inner in index]
+            index_arrays.append(inner_index)
+            index = MultiIndex.from_arrays(
+                index_arrays, names=self.index.names + [None]
+            )
 
-            index = MultiIndex.from_tuples(index, names=self.index.names + [None])
             return GeoSeries(geometries, index=index, crs=self.crs).__finalize__(self)
 
         # else PyGEOS is not available or version <= 0.8