--- a/geopandas/geoseries.py
+++ b/geopandas/geoseries.py
@@ -812,15 +812,13 @@ class GeoSeries(GeoPandasBase, Series):
             # extract original index values based on integer index
             outer_index = self.index.take(outer_idx)
 
-            index_arrays = []
-            for lvl in range(outer_index.nlevels):
-                index_arrays.append(outer_index.get_level_values(lvl).values)
-
+            nlevels = outer_index.nlevels
+            index_arrays = [outer_index.get_level_values(lvl) for lvl in range(nlevels)]
             index_arrays.append(inner_index)
+
             index = MultiIndex.from_arrays(
                 index_arrays, names=self.index.names + [None]
             )
-
             return GeoSeries(geometries, index=index, crs=self.crs).__finalize__(self)
 
         # else PyGEOS is not available or version <= 0.8