fixes #2109 

With the following setup:
```python
import numpy as np
import pandas as pd
import geopandas as gpd

outer_index = (1, 2)
nr_pts = 10000
index = pd.MultiIndex.from_arrays(
    [[outer_index] * nr_pts, np.arange(nr_pts)],
    names=("first", "second"),
)
df = gpd.GeoDataFrame(
    {"vals": np.arange(nr_pts)},
    geometry=[MultiPoint([(x, x), (x, 0)]) for x in range(nr_pts)],
    index=index,
)
```

, timing `test_df = df.explode()` I get:
`master`: `62.3 ms ± 1.06 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)`
`explode-perf`: `51.4 ms ± 1.47 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)`